:recycle: 리뷰해주신 부분 리펙터링
