+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/dto/ProfileMentorInfoDto.kt
@@ -2,7 +2,7 @@ package team.themoment.gsmNetworking.domain.mentor.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
-data class MyMentorInfoDto(
+data class ProfileMentorInfoDto(
     val id: Long,
     val name: String,
     val email: String,
@@ -11,5 +11,5 @@ data class MyMentorInfoDto(
     val sns: String?,
     val profileUrl: String?,
     val registered: Boolean,
-    val career: List<MyCareerInfoDto>
+    val career: List<ProfileCareerInfoDto>
 )
