+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentor/repository/MentorCustomRepository.kt
@@ -1,11 +1,11 @@
 package team.themoment.gsmNetworking.domain.mentor.repository
 
 import team.themoment.gsmNetworking.domain.mentor.dto.MentorInfoDto
-import team.themoment.gsmNetworking.domain.mentor.dto.MyMentorInfoDto
+import team.themoment.gsmNetworking.domain.mentor.dto.ProfileMentorInfoDto
 
 interface MentorCustomRepository {
 
     fun findAllMentorInfoDto(): List<MentorInfoDto>
 
-    fun findMyMentorInfoDto(authenticationId: Long): MyMentorInfoDto?
+    fun findMyMentorInfoDto(authenticationId: Long): ProfileMentorInfoDto?
 }
